+++ desc/objectives/_fast_ion.py
@@ -199,7 +199,7 @@ class GammaC(_Objective):
             ``Equilibrium.params_dict``.
         constants : dict
             Dictionary of constant data, e.g. transforms, profiles etc.
-            Defaults to ``self.constants``.
+            Defaults to ``self.constants``. (Deprecated)
 
         Returns
         -------
@@ -210,8 +210,7 @@ class GammaC(_Objective):
         if self._use_bounce1d:
             return self._compute_bounce1d(params, constants)
 
-        if constants is None:
-            constants = self.constants
+        constants = self._get_deprecated_constants(constants)
         eq = self.things[0]
 
         data = compute_fun(
@@ -285,8 +284,7 @@ class GammaC(_Objective):
         super().build(use_jit=use_jit, verbose=verbose)
 
     def _compute_bounce1d(self, params, constants=None):
-        if constants is None:
-            constants = self.constants
+        constants = self._get_deprecated_constants(constants)
         eq = self.things[0]
 
         data = compute_fun(