+++ desc/objectives/_generic.py
@@ -192,7 +192,7 @@ class ExternalObjective(_Objective):
         params : list of dict
             List of dictionaries of degrees of freedom, eg CoilSet.params_dict
         constants : dict
-            Unused by this Objective.
+            Unused by this Objective. (Deprecated)
 
         Returns
         -------
@@ -351,7 +351,7 @@ class GenericObjective(_Objective):
             Dictionary of equilibrium degrees of freedom, eg Equilibrium.params_dict
         constants : dict
             Dictionary of constant data, eg transforms, profiles etc.
-            Defaults to self.constants
+            Defaults to self.constants. (Deprecated)
 
         Returns
         -------
@@ -359,8 +359,7 @@ class GenericObjective(_Objective):
             Computed quantity.
 
         """
-        if constants is None:
-            constants = self.constants
+        constants = self._get_deprecated_constants(constants)
         data = compute_fun(
             self._p,
             self.f,
@@ -479,7 +478,7 @@ class LinearObjectiveFromUser(_FixedObjective):
             Dictionary of equilibrium degrees of freedom, eg thing.params_dict
         constants : dict
             Dictionary of constant data, eg transforms, profiles etc. Defaults to
-            self.constants
+            self.constants. (Deprecated)
 
         Returns
         -------
@@ -667,7 +666,7 @@ class ObjectiveFromUser(_Objective):
             Dictionary of equilibrium degrees of freedom, eg Equilibrium.params_dict
         constants : dict
             Dictionary of constant data, eg transforms, profiles etc. Defaults to
-            self.constants
+            self.constants. (Deprecated)
 
         Returns
         -------
@@ -675,8 +674,7 @@ class ObjectiveFromUser(_Objective):
             Computed quantity.
 
         """
-        if constants is None:
-            constants = self.constants
+        constants = self._get_deprecated_constants(constants)
         data = compute_fun(
             self._p,
             self._data_keys,
@@ -959,7 +957,7 @@ class DeflationOperator(_Objective):
             Dictionary of equilibrium degrees of freedom, eg Equilibrium.params_dict
         constants : dict
             Dictionary of constant data, eg transforms, profiles etc. Defaults to
-            self.constants
+            self.constants. (Deprecated)
 
         Returns
         -------