+++ desc/objectives/_geometry.py
@@ -145,7 +145,7 @@ class AspectRatio(_Objective):
             Equilibrium.params_dict
         constants : dict
             Dictionary of constant data, eg transforms, profiles etc. Defaults to
-            self.constants
+            self.constants. (Deprecated)
 
         Returns
         -------
@@ -153,8 +153,7 @@ class AspectRatio(_Objective):
             Aspect ratio, dimensionless.
 
         """
-        if constants is None:
-            constants = self.constants
+        constants = self._get_deprecated_constants(constants)
         data = compute_fun(
             self.things[0],
             self._data_keys,
@@ -291,7 +290,7 @@ class Elongation(_Objective):
             eg Equilibrium.params_dict
         constants : dict
             Dictionary of constant data, eg transforms, profiles etc. Defaults to
-            self.constants
+            self.constants. (Deprecated)
 
         Returns
         -------
@@ -299,8 +298,7 @@ class Elongation(_Objective):
             Elongation, dimensionless.
 
         """
-        if constants is None:
-            constants = self.constants
+        constants = self._get_deprecated_constants(constants)
         data = compute_fun(
             self.things[0],
             self._data_keys,
@@ -438,7 +436,7 @@ class Volume(_Objective):
             eg Equilibrium.params_dict
         constants : dict
             Dictionary of constant data, eg transforms, profiles etc. Defaults to
-            self.constants
+            self.constants. (Deprecated)
 
         Returns
         -------
@@ -446,8 +444,7 @@ class Volume(_Objective):
             Plasma volume (m^3).
 
         """
-        if constants is None:
-            constants = self.constants
+        constants = self._get_deprecated_constants(constants)
         data = compute_fun(
             self.things[0],
             self._data_keys,
@@ -741,7 +738,7 @@ class PlasmaVesselDistance(_Objective):
             Only needed if self._surface_fixed = False
         constants : dict
             Dictionary of constant data, eg transforms, profiles etc. Defaults to
-            self.constants
+            self.constants. (Deprecated)
 
         Returns
         -------
@@ -749,8 +746,7 @@ class PlasmaVesselDistance(_Objective):
             For each point in the surface grid, approximate distance to plasma.
 
         """
-        if constants is None:
-            constants = self.constants
+        constants = self._get_deprecated_constants(constants)
         if self._eq_fixed:
             surface_params = params_1
         elif self._surface_fixed:
@@ -954,7 +950,7 @@ class MeanCurvature(_Objective):
             eg Equilibrium.params_dict
         constants : dict
             Dictionary of constant data, eg transforms, profiles etc. Defaults to
-            self.constants
+            self.constants. (Deprecated)
 
         Returns
         -------
@@ -962,8 +958,7 @@ class MeanCurvature(_Objective):
             Mean curvature at each point (m^-1).
 
         """
-        if constants is None:
-            constants = self.constants
+        constants = self._get_deprecated_constants(constants)
         data = compute_fun(
             self.things[0],
             self._data_keys,
@@ -1096,7 +1091,7 @@ class PrincipalCurvature(_Objective):
             eg Equilibrium.params_dict
         constants : dict
             Dictionary of constant data, eg transforms, profiles etc. Defaults to
-            self.constants
+            self.constants. (Deprecated)
 
         Returns
         -------
@@ -1104,8 +1099,7 @@ class PrincipalCurvature(_Objective):
             Max absolute principal curvature at each point (m^-1).
 
         """
-        if constants is None:
-            constants = self.constants
+        constants = self._get_deprecated_constants(constants)
         data = compute_fun(
             self.things[0],
             self._data_keys,
@@ -1227,7 +1221,7 @@ class BScaleLength(_Objective):
             Dictionary of equilibrium degrees of freedom, eg Equilibrium.params_dict
         constants : dict
             Dictionary of constant data, eg transforms, profiles etc. Defaults to
-            self.constants
+            self.constants. (Deprecated)
 
         Returns
         -------
@@ -1235,8 +1229,7 @@ class BScaleLength(_Objective):
             Magnetic field scale length at each point (m).
 
         """
-        if constants is None:
-            constants = self.constants
+        constants = self._get_deprecated_constants(constants)
         data = compute_fun(
             "desc.equilibrium.equilibrium.Equilibrium",
             self._data_keys,
@@ -1364,7 +1357,7 @@ class GoodCoordinates(_Objective):
             Dictionary of equilibrium degrees of freedom, eg Equilibrium.params_dict
         constants : dict
             Dictionary of constant data, eg transforms, profiles etc. Defaults to
-            self.constants
+            self.constants. (Deprecated)
 
         Returns
         -------
@@ -1372,8 +1365,7 @@ class GoodCoordinates(_Objective):
             coordinate goodness error, (m^6)
 
         """
-        if constants is None:
-            constants = self.constants
+        constants = self._get_deprecated_constants(constants)
         data = compute_fun(
             "desc.equilibrium.equilibrium.Equilibrium",
             self._data_keys,
@@ -1513,7 +1505,7 @@ class MirrorRatio(_Objective):
             eg Equilibrium.params_dict
         constants : dict
             Dictionary of constant data, eg transforms, profiles etc. Defaults to
-            self.constants
+            self.constants. (Deprecated)
 
         Returns
         -------
@@ -1521,8 +1513,7 @@ class MirrorRatio(_Objective):
             Mirror ratio on each surface.
 
         """
-        if constants is None:
-            constants = self.constants
+        constants = self._get_deprecated_constants(constants)
         data = compute_fun(
             self.things[0],
             self._data_keys,