+++ desc/optimize/_constraint_wrappers.py
@@ -27,7 +27,7 @@ from desc.objectives.utils import (
     factorize_linear_constraints,
     remove_fixed_parameters,
 )
-from desc.utils import Timer, errorif, get_instance, setdefault
+from desc.utils import Timer, errorif, get_instance, setdefault, warnif
 
 from .utils import f_where_x
 
@@ -276,7 +276,7 @@ class LinearConstraintProjection(ObjectiveFunction):
         x_reduced : ndarray
             Reduced state vector that satisfies linear constraints.
         constants : list
-            Constant parameters passed to sub-objectives.
+            Constant parameters passed to sub-objectives. (Deprecated)
 
         Returns
         -------
@@ -296,7 +296,7 @@ class LinearConstraintProjection(ObjectiveFunction):
         x_reduced : ndarray
             Reduced state vector that satisfies linear constraints.
         constants : list
-            Constant parameters passed to sub-objectives.
+            Constant parameters passed to sub-objectives. (Deprecated)
 
         Returns
         -------
@@ -316,7 +316,7 @@ class LinearConstraintProjection(ObjectiveFunction):
         x_reduced : ndarray
             Reduced state vector that satisfies linear constraints.
         constants : list
-            Constant parameters passed to sub-objectives.
+            Constant parameters passed to sub-objectives. (Deprecated)
 
         Returns
         -------
@@ -336,7 +336,7 @@ class LinearConstraintProjection(ObjectiveFunction):
         x_reduced : ndarray
             Reduced state vector that satisfies linear constraints.
         constants : list
-            Constant parameters passed to sub-objectives.
+            Constant parameters passed to sub-objectives. (Deprecated)
 
         Returns
         -------
@@ -355,7 +355,7 @@ class LinearConstraintProjection(ObjectiveFunction):
         x_reduced : ndarray
             Reduced state vector that satisfies linear constraints.
         constants : list
-            Constant parameters passed to sub-objectives.
+            Constant parameters passed to sub-objectives. (Deprecated)
 
         Returns
         -------
@@ -375,7 +375,7 @@ class LinearConstraintProjection(ObjectiveFunction):
         x_reduced : ndarray
             Reduced state vector that satisfies linear constraints.
         constants : list
-            Constant parameters passed to sub-objectives.
+            Constant parameters passed to sub-objectives. (Deprecated)
 
         Returns
         -------
@@ -405,7 +405,7 @@ class LinearConstraintProjection(ObjectiveFunction):
         x_reduced : ndarray
             Reduced state vector that satisfies linear constraints.
         constants : list
-            Constant parameters passed to sub-objectives.
+            Constant parameters passed to sub-objectives. (Deprecated)
 
         Returns
         -------
@@ -423,7 +423,7 @@ class LinearConstraintProjection(ObjectiveFunction):
         x_reduced : ndarray
             Reduced state vector that satisfies linear constraints.
         constants : list
-            Constant parameters passed to sub-objectives.
+            Constant parameters passed to sub-objectives. (Deprecated)
 
         Returns
         -------
@@ -441,7 +441,7 @@ class LinearConstraintProjection(ObjectiveFunction):
         x_reduced : ndarray
             Reduced state vector that satisfies linear constraints.
         constants : list
-            Constant parameters passed to sub-objectives.
+            Constant parameters passed to sub-objectives. (Deprecated)
 
         Returns
         -------
@@ -467,7 +467,7 @@ class LinearConstraintProjection(ObjectiveFunction):
         x_reduced : ndarray
             Optimization variables with linear constraints removed.
         constants : list
-            Constant parameters passed to sub-objectives.
+            Constant parameters passed to sub-objectives. (Deprecated)
 
         """
         return self._jvp(v, x_reduced, constants, "jvp_scaled")
@@ -482,7 +482,7 @@ class LinearConstraintProjection(ObjectiveFunction):
         x_reduced : ndarray
             Optimization variables with linear constraints removed.
         constants : list
-            Constant parameters passed to sub-objectives.
+            Constant parameters passed to sub-objectives. (Deprecated)
 
         """
         return self._jvp(v, x_reduced, constants, "jvp_scaled_error")
@@ -497,7 +497,7 @@ class LinearConstraintProjection(ObjectiveFunction):
         x_reduced : ndarray
             Optimization variables with linear constraints removed.
         constants : list
-            Constant parameters passed to sub-objectives.
+            Constant parameters passed to sub-objectives. (Deprecated)
 
         """
         return self._jvp(v, x_reduced, constants, "jvp_unscaled")
@@ -517,7 +517,7 @@ class LinearConstraintProjection(ObjectiveFunction):
         x_reduced : ndarray
             Optimization variables with linear constraints removed.
         constants : list
-            Constant parameters passed to sub-objectives.
+            Constant parameters passed to sub-objectives. (Deprecated)
 
         """
         return self._vjp(v, x_reduced, constants, "vjp_scaled")
@@ -532,7 +532,7 @@ class LinearConstraintProjection(ObjectiveFunction):
         x_reduced : ndarray
             Optimization variables with linear constraints removed.
         constants : list
-            Constant parameters passed to sub-objectives.
+            Constant parameters passed to sub-objectives. (Deprecated)
 
         """
         return self._vjp(v, x_reduced, constants, "vjp_scaled_error")
@@ -547,7 +547,7 @@ class LinearConstraintProjection(ObjectiveFunction):
         x_reduced : ndarray
             Optimization variables with linear constraints removed.
         constants : list
-            Constant parameters passed to sub-objectives.
+            Constant parameters passed to sub-objectives. (Deprecated)
 
         """
         return self._vjp(v, x_reduced, constants, "vjp_unscaled")
@@ -976,7 +976,7 @@ class ProximalProjection(ObjectiveFunction):
         x : ndarray
             State vector.
         constants : list
-            Constant parameters passed to sub-objectives.
+            Constant parameters passed to sub-objectives. (Deprecated)
 
         Returns
         -------
@@ -984,7 +984,7 @@ class ProximalProjection(ObjectiveFunction):
             Objective function value(s).
 
         """
-        constants = setdefault(constants, self.constants)
+        constants = setdefault(constants, [None, None])
         xopt, _ = self._update_equilibrium(x, store=False)
         return self._objective.compute_scaled(xopt, constants[0])
 
@@ -996,7 +996,7 @@ class ProximalProjection(ObjectiveFunction):
         x : ndarray
             State vector.
         constants : list
-            Constant parameters passed to sub-objectives.
+            Constant parameters passed to sub-objectives. (Deprecated)
 
         Returns
         -------
@@ -1004,7 +1004,7 @@ class ProximalProjection(ObjectiveFunction):
             Objective function value(s).
 
         """
-        constants = setdefault(constants, self.constants)
+        constants = setdefault(constants, [None, None])
         xopt, _ = self._update_equilibrium(x, store=False)
         return self._objective.compute_scaled_error(xopt, constants[0])
 
@@ -1016,7 +1016,7 @@ class ProximalProjection(ObjectiveFunction):
         x : ndarray
             State vector.
         constants : list
-            Constant parameters passed to sub-objectives.
+            Constant parameters passed to sub-objectives. (Deprecated)
 
         Returns
         -------
@@ -1035,7 +1035,7 @@ class ProximalProjection(ObjectiveFunction):
         x : ndarray
             State vector.
         constants : list
-            Constant parameters passed to sub-objectives.
+            Constant parameters passed to sub-objectives. (Deprecated)
 
         Returns
         -------
@@ -1043,7 +1043,7 @@ class ProximalProjection(ObjectiveFunction):
             Objective function value(s).
 
         """
-        constants = setdefault(constants, self.constants)
+        constants = setdefault(constants, [None, None])
         xopt, _ = self._update_equilibrium(x, store=False)
         return self._objective.compute_unscaled(xopt, constants[0])
 
@@ -1055,7 +1055,7 @@ class ProximalProjection(ObjectiveFunction):
         x : ndarray
             State vector.
         constants : list
-            Constant parameters passed to sub-objectives.
+            Constant parameters passed to sub-objectives. (Deprecated)
 
         Returns
         -------
@@ -1073,7 +1073,7 @@ class ProximalProjection(ObjectiveFunction):
         # ∇L = G.T @ ∇G @ [dc_tangents - (∇F @ dx_tangents) ^ -1 @ (∇F @ dc_tangents)]
         # We get the part in [] using the _get_tangent method.
         v = jnp.eye(x.shape[0])
-        constants = setdefault(constants, self.constants)
+        constants = setdefault(constants, [None, None])
         xg, xf = self._update_equilibrium(x, store=True)
         if not (self._constraint._is_mpi or self._objective._is_mpi):
             jvpfun = lambda u: self._get_tangent(u, xf, constants, op="scaled_error")
@@ -1108,7 +1108,7 @@ class ProximalProjection(ObjectiveFunction):
         x : ndarray
             State vector.
         constants : list
-            Constant parameters passed to sub-objectives.
+            Constant parameters passed to sub-objectives. (Deprecated)
 
         Returns
         -------
@@ -1127,7 +1127,7 @@ class ProximalProjection(ObjectiveFunction):
         x : ndarray
             State vector.
         constants : list
-            Constant parameters passed to sub-objectives.
+            Constant parameters passed to sub-objectives. (Deprecated)
 
         Returns
         -------
@@ -1146,7 +1146,7 @@ class ProximalProjection(ObjectiveFunction):
         x : ndarray
             State vector.
         constants : list
-            Constant parameters passed to sub-objectives.
+            Constant parameters passed to sub-objectives. (Deprecated)
 
         Returns
         -------
@@ -1165,7 +1165,7 @@ class ProximalProjection(ObjectiveFunction):
         x : ndarray
             State vector.
         constants : list
-            Constant parameters passed to sub-objectives.
+            Constant parameters passed to sub-objectives. (Deprecated)
 
         Returns
         -------
@@ -1186,7 +1186,7 @@ class ProximalProjection(ObjectiveFunction):
         x : ndarray
             Optimization variables.
         constants : list
-            Constant parameters passed to sub-objectives.
+            Constant parameters passed to sub-objectives. (Deprecated)
 
         """
         op = "scaled"
@@ -1203,7 +1203,7 @@ class ProximalProjection(ObjectiveFunction):
         x : ndarray
             Optimization variables.
         constants : list
-            Constant parameters passed to sub-objectives.
+            Constant parameters passed to sub-objectives. (Deprecated)
 
         """
         op = "scaled_error"
@@ -1220,7 +1220,7 @@ class ProximalProjection(ObjectiveFunction):
         x : ndarray
             Optimization variables.
         constants : list
-            Constant parameters passed to sub-objectives.
+            Constant parameters passed to sub-objectives. (Deprecated)
 
         """
         op = "unscaled"
@@ -1243,7 +1243,7 @@ class ProximalProjection(ObjectiveFunction):
         # First we will compute the tangent direction (see _get_tangent for details),
         # then we will compute the Jacobian.
         v = v[0] if isinstance(v, (tuple, list)) else v
-        constants = setdefault(constants, self.constants)
+        constants = setdefault(constants, [None, None])
         xg, xf = self._update_equilibrium(x, store=True)
 
         # we don't need to divide this part into blocked and batched because
@@ -1330,6 +1330,14 @@ class ProximalProjection(ObjectiveFunction):
     @property
     def constants(self):
         """list: constant parameters for each sub-objective."""
+        warnif(
+            True,
+            FutureWarning,
+            "constants is deprecated and will be removed in a future "
+            "release. Users should not include constants in the arguments "
+            "of their objective compute methods. Instead declare all the "
+            "constants in the build method and use as obj._constants.",
+        )
         return [self._objective.constants, self._constraint.constants]
 
     def __getattr__(self, name):
@@ -1397,7 +1405,7 @@ def _proximal_jvp_blocked_pure(objective, vgs, xgs, op):
     # Note: This function is very similar to _jvp_blocked in ObjectiveFunction with
     # some naming differences to account for ProximalProjection.
     out = []
-    for k, (obj, const) in enumerate(zip(objective.objectives, objective.constants)):
+    for k, obj in enumerate(objective.objectives):
         thing_idx = objective._things_per_objective_idx[k]
         xi = [xgs[i] for i in thing_idx]
         vi = [vgs[i] for i in thing_idx]
@@ -1408,11 +1416,11 @@ def _proximal_jvp_blocked_pure(objective, vgs, xgs, op):
             # obj might not allow fwd mode, so compute full rev mode jacobian
             # and do matmul manually. This is slightly inefficient, but usually
             # when rev mode is used, dim_f <<< dim_x, so its not too bad.
-            Ji = getattr(obj, "jac_" + op)(*xi, constants=const)
+            Ji = getattr(obj, "jac_" + op)(*xi)
             outi = jnp.array([Jii @ vii.T for Jii, vii in zip(Ji, vi)]).sum(axis=0)
             out.append(outi)
         else:
-            outi = getattr(obj, "jvp_" + op)([_vi for _vi in vi], xi, constants=const).T
+            outi = getattr(obj, "jvp_" + op)([_vi for _vi in vi], xi).T
             out.append(outi)
 
     return jnp.concatenate(out).T